[Feat] #205 - 데이터 바인딩 옵셔널로 처리

+++ HealthFoodMe/HealthFoodMe/Presentation/Detail/MenuTabScene/ViewModel/MenuDataModel.swift
@@ -13,7 +13,7 @@ struct MenuDataModel: Codable {
     let memuImageURL: String?
     let menuName: String
     let menuPrice: Int
-    let menuKcal: Int?
+    let menuKcal: Double?
     let carbohydrates: Int?
     let protein: Int?
     let fat: Int?
